game/pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "generals-bots>=2.5.0",
]
